# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProFile API",
    "description": "ProFile Resume & Portfolio Management API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/resumes/manage/user/{userId}": {
      "get": {
        "operationId": "resumes_listResumesForUser",
        "tags": [
          "resumes"
        ],
        "summary": "List all resumes for a specific user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all resumes for a specific user"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/onboarding/status": {
      "get": {
        "operationId": "onboarding_getStatus",
        "tags": [
          "onboarding"
        ],
        "summary": "Get user onboarding status",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get user onboarding status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingProgressResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/onboarding/progress": {
      "get": {
        "operationId": "onboarding_getProgress",
        "tags": [
          "onboarding"
        ],
        "summary": "Get user onboarding progress (checkpoint)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get user onboarding progress (checkpoint)"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "put": {
        "operationId": "onboarding_saveProgress",
        "tags": [
          "onboarding"
        ],
        "summary": "Save user onboarding progress (checkpoint)",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Save user onboarding progress (checkpoint)"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/onboarding": {
      "post": {
        "operationId": "onboarding_completeOnboarding",
        "tags": [
          "onboarding"
        ],
        "summary": "Complete user onboarding with resume data",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Complete user onboarding with resume data"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/themes/approval/pending": {
      "get": {
        "operationId": "themes_getPending",
        "tags": [
          "themes"
        ],
        "summary": "Get pending themes for approval",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get pending themes for approval"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/approval/review": {
      "post": {
        "operationId": "themes_review",
        "tags": [
          "themes"
        ],
        "summary": "Review and approve/reject a theme",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Review and approve/reject a theme"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/approval/{id}/submit": {
      "post": {
        "operationId": "themes_submit",
        "tags": [
          "themes"
        ],
        "summary": "Submit a theme for approval",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Submit a theme for approval"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/me": {
      "get": {
        "operationId": "themes_getAllThemesByUser",
        "tags": [
          "themes"
        ],
        "summary": "Get my themes",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get my themes"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/resume/{resumeId}/config": {
      "get": {
        "operationId": "themes_getResolvedConfig",
        "tags": [
          "themes"
        ],
        "summary": "Get resolved config for resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get resolved config for resume"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes": {
      "post": {
        "operationId": "themes_createThemeForUser",
        "tags": [
          "themes"
        ],
        "summary": "Create theme",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Create theme"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "get": {
        "operationId": "themes_findAllThemesWithPagination",
        "tags": [
          "themes"
        ],
        "summary": "List published themes",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List published themes"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/fork": {
      "post": {
        "operationId": "themes_fork",
        "tags": [
          "themes"
        ],
        "summary": "Fork a theme",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fork a theme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/apply": {
      "post": {
        "operationId": "themes_apply",
        "tags": [
          "themes"
        ],
        "summary": "Apply theme to resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Apply theme to resume"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/themes/{id}": {
      "put": {
        "operationId": "themes_updateThemeForUser",
        "tags": [
          "themes"
        ],
        "summary": "Update theme",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update theme"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "delete": {
        "operationId": "themes_deleteThemeForUser",
        "tags": [
          "themes"
        ],
        "summary": "Delete theme",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete theme"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "get": {
        "operationId": "themes_findThemeById",
        "tags": [
          "themes"
        ],
        "summary": "Get theme by ID",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get theme by ID"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OnboardingProgressResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "clxxx..."
          },
          "currentStep": {
            "type": "number",
            "example": "3"
          },
          "totalSteps": {
            "type": "number",
            "example": "5"
          },
          "progressPercentage": {
            "type": "number",
            "example": "60"
          },
          "completed": {
            "type": "boolean",
            "example": "false"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardingStepResponseDto"
            }
          }
        },
        "required": [
          "userId",
          "currentStep",
          "totalSteps",
          "progressPercentage",
          "completed",
          "steps"
        ]
      },
...
      "ThemeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "name": {
            "type": "string"
          },
          "isSystem": {
            "type": "boolean",
            "example": "false"
          },
          "isPublic": {
            "type": "boolean",
            "example": "false"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "isSystem",
          "isPublic",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "OnboardingStepResponseDto": {
        "type": "object",
        "properties": {
          "stepId": {
            "type": "string",
            "example": "PROFILE_SETUP"
          },
          "title": {
            "type": "string"
          },
          "completed": {
            "type": "boolean",
            "example": "true"
          },
          "description": {
            "type": "string"
          },
          "completedAt": {
            "type": "string"
          }
        },
        "required": [
          "stepId",
          "title",
          "completed"
        ]
      },
...
    }
  }
}